Capitalize IP acronym
